マルチスレッド改善

+++ ZoppaLogger/ZoppaLogger/Logger.vb
@@ -3,39 +3,45 @@ Option Explicit On
 
 Imports System.IO
 Imports System.IO.Compression
-Imports System.Reflection
 Imports System.Runtime.CompilerServices
 Imports System.Threading
 
 ''' <summary>ログ出力機能。</summary>
-Public NotInheritable Class Logger
+Public Class Logger
 
-    ' 単一インスタンス
-    Private Shared mLogger As Logger
+    ''' <summary>カレントのログファイルが圧縮されたことを通知します。</summary>
+    ''' <param name="sender">イベント発行元。</param>
+    ''' <param name="e">イベントオブジェクト。</param>
+    Public Event NotificationCompressedFile(sender As Object, e As NotificationCompressedFileEventArgs)
+
+    ''' <summary>圧縮済みログファイルが最大世代数を超えたことを通知します。</summary>
+    ''' <param name="sender">イベント発行元。</param>
+    ''' <param name="e">イベントオブジェクト。</param>
+    Public Event NotificationOrganizeCompressedFile(sender As Object, e As NotificationOrganizeCompressedFileEventArgs)
 
     ' 対象ファイル
-    Private ReadOnly mLogFile As FileInfo
+    Private mLogFile As FileInfo
 
     ' 出力エンコード
-    Private ReadOnly mEncode As Text.Encoding
+    Private mEncode As Text.Encoding
 
     ' 最大ログサイズ
-    Private ReadOnly mMaxLogSize As Integer
+    Private mMaxLogSize As Integer
 
     ' 最大ログ世代数
-    Private ReadOnly mLogGen As Integer
+    Private mLogGen As Integer
 
     ' ログ出力レベル
     Private mLogLevel As LogLevel
 
     ' 日付が変わったら切り替えるかのフラグ
-    Private ReadOnly mDateChange As Boolean
+    Private mDateChange As Boolean
 
     ' キャッシュに保存するログ行数のリミット
-    Private ReadOnly mCacheLimit As Integer
+    Private mCacheLimit As Integer
 
     ' ログ出力書式メソッド
-    Private ReadOnly mFormatMethod As Func(Of LogData, String)
+    Private mFormatMethod As Func(Of LogData, String)
 
     ' 書込みバッファ
     Private ReadOnly mQueue As New Queue(Of LogData)()
@@ -45,35 +51,6 @@ Public NotInheritable Class Logger
 
     ' 書込み中フラグ
     Private mWriting As Boolean
-
-    ''' <summary>ログ設定を行う。</summary>
-    ''' <param name="logFilePath">出力ファイル名。</param>
-    ''' <param name="encode">出力エンコード。</param>
-    ''' <param name="maxLogSize">最大ログファイルサイズ。</param>
-    ''' <param name="logGeneration">ログ世代数。</param>
-    ''' <param name="logLevel">ログ出力レベル。</param>
-    ''' <param name="dateChange">日付の変更でログを切り替えるかの設定。</param>
-    ''' <param name="cacheLimit">ログを貯めて置くリミット（超えたらログ出力を優先）</param>
-    ''' <param name="formatMethod">ログ出力書式メソッド。</param>
-    Private Sub New(logFilePath As String,
-                    encode As Text.Encoding,
-                    maxLogSize As Integer,
-                    logGeneration As Integer,
-                    logLevel As LogLevel,
-                    dateChange As Boolean,
-                    cacheLimit As Integer,
-                    formatMethod As Func(Of LogData, String))
-        Me.mLogFile = New FileInfo(logFilePath)
-        Me.mEncode = encode
-        Me.mMaxLogSize = maxLogSize
-        Me.mLogGen = logGeneration
-        Me.mWriting = False
-        Me.mLogLevel = logLevel
-        Me.mDateChange = dateChange
-        Me.mPrevWriteDate = Date.MaxValue
-        Me.mCacheLimit = cacheLimit
-        Me.mFormatMethod = formatMethod
-    End Sub
 
     ''' <summary>デフォルトログを使用します。</summary>
     ''' <param name="logFilePath">出力ファイルパス。</param>
@@ -92,18 +69,45 @@ Public NotInheritable Class Logger
                                Optional dateChange As Boolean = False,
                                Optional cacheLimit As Integer = 1000,
                                Optional formatMethod As Func(Of LogData, String) = Nothing) As Logger
-        mLogger?.WaitFinish()
+        Return UseCustom(Of Logger)(logFilePath, encode, maxLogSize, logGeneration, logLevel, dateChange, cacheLimit, formatMethod)
+    End Function
 
+    ''' <summary>カスタムログを使用します。</summary>
+    ''' <param name="logFilePath">出力ファイルパス。</param>
+    ''' <param name="encode">出力エンコード。</param>
+    ''' <param name="maxLogSize">最大ログファイルサイズ。</param>
+    ''' <param name="logGeneration">ログ世代数。</param>
+    ''' <param name="logLevel">ログレベル。</param>
+    ''' <param name="dateChange">日付の変更でログを切り替えるかの設定。</param>
+    ''' <param name="cacheLimit">ログを貯めて置くリミット（超えたらログ出力を優先）</param>
+    ''' <param name="formatMethod">ログ出力書式メソッド。</param>
+    Public Shared Function UseCustom(Of T As {Logger, New})(Optional logFilePath As String = "default.log",
+                                                            Optional encode As Text.Encoding = Nothing,
+                                                            Optional maxLogSize As Integer = 30 * 1024 * 1024,
+                                                            Optional logGeneration As Integer = 10,
+                                                            Optional logLevel As LogLevel = LogLevel.Debug,
+                                                            Optional dateChange As Boolean = False,
+                                                            Optional cacheLimit As Integer = 1000,
+                                                            Optional formatMethod As Func(Of LogData, String) = Nothing) As T
         Dim fi As New FileInfo(logFilePath)
         If Not fi.Directory.Exists Then
             fi.Directory.Create()
         End If
-        mLogger = New Logger(
-            logFilePath, If(encode, Text.Encoding.Default),
-            maxLogSize, logGeneration, logLevel, dateChange, cacheLimit,
-            If(formatMethod, New Func(Of LogData, String)(AddressOf LogFormat))
-        )
-        Return mLogger
+        Dim res = New T()
+        With res
+            .mLogFile = New FileInfo(logFilePath)
+            .mEncode = If(encode, Text.Encoding.Default)
+            .mMaxLogSize = maxLogSize
+            .mLogGen = logGeneration
+            .mLogLevel = logLevel
+            .mDateChange = dateChange
+            .mCacheLimit = cacheLimit
+            .mFormatMethod = If(formatMethod, New Func(Of LogData, String)(AddressOf res.OnFormatting))
+
+            .mWriting = False
+            .mPrevWriteDate = Date.MaxValue
+        End With
+        Return res
     End Function
 
     ''' <summary>ログをファイルに出力します。</summary>
@@ -148,16 +152,18 @@ Public NotInheritable Class Logger
         If Me.mLogFile.Exists AndAlso
                (Me.mLogFile.Length > Me.mMaxLogSize OrElse Me.ChangeOfDate) Then
             Try
-                ' 以前のファイルをリネーム
+                ' ファイル名の要素を分割
                 Dim ext = Path.GetExtension(Me.mLogFile.Name)
                 Dim nm = Me.mLogFile.Name.Substring(0, Me.mLogFile.Name.Length - ext.Length)
                 Dim tn = Date.Now.ToString("yyyyMMddHHmmssfff")
 
+                ' 圧縮するフォルダを作成
                 Dim zipPath = New IO.FileInfo($"{mLogFile.Directory.FullName}\{nm}_{tn}\{nm}{ext}")
                 If Not zipPath.Exists Then
                     zipPath.Directory.Create()
                 End If
                 Try
+                    ' ログファイルを圧縮するフォルダへ移動
                     Dim moved = False
                     Dim exx As Exception = Nothing
                     For i As Integer = 0 To 4
@@ -168,12 +174,27 @@ Public NotInheritable Class Logger
                         Catch ex As Exception
                             exx = ex
                             Thread.Sleep(100)
+                            Debug.WriteLine($"Write 1 {ex.Message}")
                         End Try
                     Next
+
+                    ' 移動できたらログファイルを圧縮
                     If moved Then
-                        ZipFile.CreateFromDirectory(
-                                zipPath.Directory.FullName, $"{zipPath.Directory.FullName}.zip"
-                            )
+                        Dim compressFile = $"{zipPath.Directory.FullName}.zip"
+
+                        ' 現在のログファイルを圧縮
+                        ZipFile.CreateFromDirectory(zipPath.Directory.FullName, compressFile)
+
+                        ' ログファイルを圧縮したことを外部に通知
+                        Task.Run(
+                            Sub()
+                                Try
+                                    Me.OnNotificationCompressedFile(New NotificationCompressedFileEventArgs(New FileInfo(compressFile)))
+                                Catch ex As Exception
+                                    Debug.WriteLine($"Write 2 {ex.Message}")
+                                End Try
+                            End Sub
+                        )
                     Else
                         Throw exx
                     End If
@@ -185,36 +206,59 @@ Public NotInheritable Class Logger
 
                 ' 過去ファイルを整理
                 Dim oldfiles = Directory.GetFiles(Me.mLogFile.Directory.FullName, $"{nm}*.zip").ToList()
-                oldfiles.Sort()
-                Do While oldfiles.Count > Me.mLogGen
-                    File.Delete(oldfiles.First())
-                    oldfiles.RemoveAt(0)
-                Loop
+                If oldfiles.Count > Me.mLogGen Then
+                    Task.Run(
+                        Sub()
+                            ' 削除順にファイルをソート
+                            oldfiles.Sort()
+
+                            ' 削除するファイルを外部に通知
+                            Dim ev As New NotificationOrganizeCompressedFileEventArgs(oldfiles)
+                            Me.OnNotificationOrganizeCompressedFile(ev)
+
+                            ' キャンセルされていなければ削除
+                            If Not ev.Cancel Then
+                                Do While oldfiles.Count > Me.mLogGen
+                                    If File.Exists(oldfiles.First()) Then
+                                        File.Delete(oldfiles.First())
+                                        oldfiles.RemoveAt(0)
+                                    End If
+                                Loop
+                            End If
+                        End Sub
+                    )
+                End If
 
             Catch ex As Exception
                 SyncLock Me
                     Me.mWriting = False
                 End SyncLock
+                Debug.WriteLine($"Write 3 {ex.Message}")
                 Return
             End Try
         End If
 
         Try
+            Me.mLogFile.Refresh()
             Using sw As New StreamWriter(Me.mLogFile.FullName, True, Me.mEncode)
                 Dim writed As Boolean
                 Do
                     ' キュー内の文字列を取得
+                    '
+                    ' 1. キューにログ情報がある
+                    '    対象ログレベル以上のログレベルを出力する場合、出力する
+                    ' 2. キューにログ情報が空の場合はループを抜けてファイルストリームを閉じる
                     writed = False
                     Dim ln As LogData? = Nothing
                     Dim outd As Boolean = False
                     SyncLock Me
                         If Me.mQueue.Count > 0 Then
-                            ln = Me.mQueue.Dequeue()
+                            ln = Me.mQueue.Dequeue()                    ' 1
                             If Me.mLogLevel >= ln.Value.LogLevel Then
                                 outd = True
                             End If
                         Else
-                            Me.mWriting = False
+                            Exit Do                                     ' 2
                         End If
                     End SyncLock
 
@@ -224,15 +268,18 @@ Public NotInheritable Class Logger
                         writed = True
                     End If
 
+                    ' 出力した結果、ログファイルが最大サイズを超える場合、ループを抜けてストリームを閉じる
                     Me.mLogFile.Refresh()
                     If Me.mLogFile.Length > Me.mMaxLogSize OrElse Me.ChangeOfDate Then
-                        SyncLock Me
-                            Me.mWriting = False
-                        End SyncLock
-                        Return
+                        Exit Do
                     End If
                 Loop While writed
             End Using
+
+            ' 上のループを抜けたとき実行中フラグを落とす
+            SyncLock Me
+                Me.mWriting = False
+            End SyncLock
 
             Threading.Thread.Sleep(10)
 
@@ -240,12 +287,15 @@ Public NotInheritable Class Logger
             SyncLock Me
                 Me.mWriting = False
             End SyncLock
+            Debug.WriteLine($"Write 4 {ex.Message}")
         Finally
             Me.mPrevWriteDate = Date.Now
         End Try
     End Sub
 
-    Private Shared Function LogFormat(dat As LogData) As String
+    ''' <summary>ログをファイルに出力する文字列を作成する。</summary>
+    ''' <param name="dat">ログデータ。</param>
+    Protected Overridable Function OnFormatting(dat As LogData) As String
         Dim lv = "INFO "
         Select Case dat.LogLevel
             Case LogLevel.Debug
@@ -260,6 +310,18 @@ Public NotInheritable Class Logger
 
         Return $"[{dat.WriteTime:yyyy/MM/dd HH:mm:ss} {lv} {If(dat.CallerClass IsNot Nothing, $"{dat.CallerClass.Name}.{dat.CallerMethod}({dat.LineNo})", "")}] {dat.LogMessage}"
     End Function
+
+    ''' <summary>カレントのログファイルが圧縮されたイベントを発行します。</summary>
+    ''' <param name="e">イベントオブジェクト。</param>
+    Protected Overridable Sub OnNotificationCompressedFile(e As NotificationCompressedFileEventArgs)
+        RaiseEvent NotificationCompressedFile(Me, e)
+    End Sub
+
+    ''' <summary>圧縮済みログファイルが最大世代数を超えたイベントを発行します。</summary>
+    ''' <param name="e">イベントオブジェクト。</param>
+    Protected Overridable Sub OnNotificationOrganizeCompressedFile(e As NotificationOrganizeCompressedFileEventArgs)
+        RaiseEvent NotificationOrganizeCompressedFile(Me, e)
+    End Sub
 
     ''' <summary>ログレベルを設定します。</summary>
     ''' <param name="lv">新しいログレベル。</param>
@@ -357,7 +419,7 @@ Public NotInheritable Class Logger
             End If
 
         Catch ex As Exception
-
+            Debug.WriteLine($"FlushWrite {ex.Message}")
         End Try
     End Sub
 
@@ -418,5 +480,40 @@ Public NotInheritable Class Logger
         End Sub
 
     End Structure
+
+    ''' <summary>ログファイル圧縮通知イベント情報。</summary>
+    Public NotInheritable Class NotificationCompressedFileEventArgs
+        Inherits EventArgs
+
+        ''' <summary>圧縮したログファイルを示す情報を取得します。</summary>
+        Public ReadOnly Property TargetFile As IO.FileInfo
+
+        ''' <summary>コンストラクタ。</summary>
+        ''' <param name="targetFile">圧縮したログファイル。</param>
+        Public Sub New(targetFile As IO.FileInfo)
+            Me.TargetFile = targetFile
+        End Sub
+
+    End Class
+
+    ''' <summary>圧縮ログファイルの整理通知イベント情報。</summary>
+    Public NotInheritable Class NotificationOrganizeCompressedFileEventArgs
+        Inherits EventArgs
+
+        ''' <summary>圧縮したログファイルのリストを取得します。</summary>
+        Public ReadOnly Property TargetFiles As New List(Of IO.FileInfo)
+
+        ''' <summary>古いファイルを削除のキャンセルを設定、取得します。</summary>
+        Public Property Cancel As Boolean = False
+
+        ''' <summary>コンストラクタ。</summary>
+        ''' <param name="targetPaths">圧縮したログファイルリスト。</param>
+        Public Sub New(targetPaths As List(Of String))
+            For Each ph In targetPaths
+                Me.TargetFiles.Add(New IO.FileInfo(ph))
+            Next
+        End Sub
+
+    End Class
 
 End Class
